+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Splunk Python SDK Changelog
 
+## 0.8.5
+
+* Collection.__iter__ now iterates over names of entities, not the
+  entities themselves, for consistency with a dict-list interface.
+  Collection also support len, del, .keys(), .iteritems(),
+  .itervalues(), and .iterkeys().
+	
 ## 0.8.0 (beta)
 
 ### Features